requirements.txt
# Telegram bot için gerekli kütüphane
python-telegram-bot==20.6

# HTTP istekleri için
requests==2.31.0

# .env dosyasını okumak için
python-dotenv==1.0.0

# Temel işlemler için
pandas==2.1.0
numpy>=1.26.0

# Web scraping için (OBS sistemine bağlanmak için gerekebilir)
beautifulsoup4==4.12.2
lxml>=5.1.0

# Güvenlik için
cryptography>=42.0.0 

# Zamanlama için
schedule>=1.2.1

# GUI için
PyQt5==5.15.9

